fix: import shared types from the index
this means the dts bundle won't have duplicates, which is important
for the unique symbol.
