+++ src/shared.ts
@@ -1,10 +1,7 @@
-interface ToString {
-	toString(): string
-}
+import { html, type Displayable, type HTML, type Renderable } from 'dhtml'
 
-export type Displayable = null | undefined | ToString | Node | Renderable | Iterable<Displayable>
-export interface Renderable {
-	render(): Displayable
+declare global {
+	var __DEV__: boolean
 }
 
 export const is_renderable = (value: unknown): value is Renderable =>
@@ -13,31 +10,12 @@ export const is_renderable = (value: unknown): value is Renderable =>
 export const is_iterable = (value: unknown): value is Iterable<unknown> =>
 	typeof value === 'object' && value !== null && Symbol.iterator in value
 
-declare global {
-	var __DEV__: boolean
-}
-
 export function assert(value: unknown, message?: string): asserts value {
 	if (!__DEV__) return
 	if (!value) throw new Error(message ?? 'assertion failed')
 }
 
-const tag: unique symbol = Symbol()
-
-interface HTML {
-	[tag]: true
-	/* @internal */ _statics: TemplateStringsArray
-	/* @internal */ _dynamics: unknown[]
-}
-
-export const is_html = (value: any): value is HTML => typeof value === 'object' && value !== null && tag in value
-
-export function html(statics: TemplateStringsArray, ...dynamics: unknown[]): HTML {
-	return {
-		[tag]: true,
-		_dynamics: dynamics,
-		_statics: statics,
-	}
-}
+export const html_tag: unique symbol = Symbol()
+export const is_html = (value: any): value is HTML => typeof value === 'object' && value !== null && html_tag in value
 
 export const single_part_template = (part: Displayable): HTML => html`${part}`
